requirements.txt
hydrogram
tgcrypto
pymongo
cinemagoer
shortzy
aiohttp
aiofiles
speedtest-cli
uvloop
requests
qrcode
Pillow
pytz
motor
umongo
dnspython
